requirements.txt
asgiref==3.2.3
Django==3.0.3
django-autoslug==1.9.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
greenlet==0.4.15
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
Unidecode==0.4.19
